refactor(sandbox_hints): split into module hierarchy by detection strategy
