+++ clash/src/cmd/hooks.rs
@@ -155,7 +155,7 @@ impl HooksCmd {
                         crate::network_hints::check_for_sandbox_network_hint(&input, s)
                     });
                     let fs = settings.as_ref().and_then(|s| {
-                        crate::sandbox_fs_hints::check_for_sandbox_fs_hint(&input, s)
+                        crate::sandbox_hints::check_for_sandbox_fs_hint(&input, s)
                     });
                     (net, fs)
                 };
