+++ clash/src/lib.rs
@@ -47,7 +47,7 @@ pub mod policy;
 pub mod policy_loader;
 pub mod sandbox;
 pub mod sandbox_cmd;
-pub mod sandbox_fs_hints;
+pub mod sandbox_hints;
 pub mod schema;
 pub mod session_policy;
 pub mod settings;
